+++ libs/ngx-pass-code/CHANGELOG.md
@@ -4,4 +4,9 @@ Initial release. Unpublished.
 
 ### 1.0.1 - 06.12.2022.
 
-Official release.
+Official release. Unpublished.
+
+### 1.0.2 - 06.12.2022.
+
+Bugfix on setDisabledState method on initial assigned control state. Fix value
+emit on set/patch value.